Move to beneficiary hub package
