+++ form/src/main/java/org/dhis2/form/ui/beneficiaryHub/DateOfBirthFormatter.kt
@@ -1,4 +1,4 @@
-package org.dhis2.form.utils
+package org.dhis2.form.ui.beneficiaryHub
 
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
