+++ form/src/test/java/org/dhis2/form/beneficiaryHub/AgeCalculatorTest.kt
@@ -1,5 +1,6 @@
-package org.dhis2.form.utils
+package org.dhis2.form.beneficiaryHub
 
+import org.dhis2.form.ui.beneficiaryHub.AgeCalculator
 import org.junit.Assert.assertEquals
 import org.junit.Test
 import java.time.Clock
